requirements.txt
absl-py==0.13.0
cleverhans==4.0.0
cloudpickle==2.0.0
cycler==0.10.0
dataclasses==0.8
decorator==5.1.0
dm-tree==0.1.6
easydict==1.9
gast==0.5.2
joblib==1.0.1
kiwisolver==1.3.1
matplotlib==3.3.4
mnist==0.2.2
nose==1.3.7
numpy==1.19.5
Pillow==8.3.2
pycodestyle==2.7.0
pyparsing==2.4.7
python-dateutil==2.8.2
scipy==1.5.4
six==1.16.0
tensorflow-probability==0.14.0
torch==1.9.0+cu102
torchaudio==0.9.0
torchvision==0.10.0+cu102
typing-extensions==3.10.0.2
